requirements.txt
pytorch==1.11.0
torchvision==0.12.0
addict
future
lmdb
numpy
opencv-python
Pillow
pyyaml
requests
scikit-image
scipy
tb-nightly
tqdm
yapf
